pipeline/requirements.txt
scenedetect[opencv]
google-genai
anthropic
httpx
python-dotenv
ffmpeg-python
psycopg2-binary
